Do not assume the first found window in [[UIApplication sharedApplication] windows] to be the right one
Before showing the alert view it is added on the first object found in [[UIApplication sharedApplication] windows].
Starting with iOS 9 Apple introduced a feature called Picture in Picture; when this feature is enabled
the first found window will be an object of type `PGHostedWindow` which holds the popup view of the PiP feature.
Trying to add the alert view on that window will make the alert view to not appear at all.